be/django/requirements-2.txt
colorama==0.3.3
Django==1.8.2
djangorestframework==3.1.3
green==1.11.0
mock==1.0.1
python-termstyle==0.1.10
requests==2.7.0
six==1.9.0
wheel==0.24.0
